+++ BatB/app/build.gradle
@@ -34,4 +34,5 @@ dependencies {
     implementation 'com.github.devlight.navigationtabstrip:navigationtabstrip:1.0.4'
     implementation 'com.github.bumptech.glide:glide:4.8.0'
     implementation 'org.jetbrains:annotations-java5:15.0'
+    implementation "android.arch.lifecycle:extensions:1.1.1"
 }
