+++ BatB/app/src/main/AndroidManifest.xml
@@ -9,6 +9,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
+        <activity
+            android:name="listSplashActivity"
+            android:theme="@android:style/Theme.NoTitleBar.Fullscreen" >
+        </activity>
         <activity android:name=".HelpActivity"></activity>
         <activity android:name=".MainActivity">
             <intent-filter>
