--- BatB/app/src/main/java/com/example/batb/HorizontalPagerFragment.java
@@ -1,49 +1,0 @@
-package com.example.batb;
-
-import android.content.Context;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.example.batb.adapters.HorizontalPagerAdapter;
-import com.gigamole.infinitecycleviewpager.HorizontalInfiniteCycleViewPager;
-
-public class HorizontalPagerFragment extends Fragment {
-    public static int celebid;
-    @Nullable
-    @Override
-    public View onCreateView(final LayoutInflater inflater, @Nullable final ViewGroup container, @Nullable final Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_horizontal, container, false);
-    }
-    public static int getID(int s){
-        return celebid = s;
-    }
-
-    @Override
-    public void onViewCreated(final View view, @Nullable final Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        final HorizontalInfiniteCycleViewPager horizontalInfiniteCycleViewPager =
-                (HorizontalInfiniteCycleViewPager) view.findViewById(R.id.hicvp);
-        horizontalInfiniteCycleViewPager.setAdapter(new HorizontalPagerAdapter(getContext(), celebid));
-
-//        horizontalInfiniteCycleViewPager.setScrollDuration(400);
-//        horizontalInfiniteCycleViewPager.setPageDuration(1000);
-//        horizontalInfiniteCycleViewPager.setInterpolator(
-//                AnimationUtils.loadInterpolator(getContext(), android.R.anim.overshoot_interpolator)
-//        );
-//        horizontalInfiniteCycleViewPager.setMediumScaled(false);
-//        horizontalInfiniteCycleViewPager.setMaxPageScale(0.8F);
-//        horizontalInfiniteCycleViewPager.setMinPageScale(0.5F);
-//        horizontalInfiniteCycleViewPager.setCenterPageScaleOffset(30.0F);
-//        horizontalInfiniteCycleViewPager.setMinPageScaleOffset(5.0F);
-//        horizontalInfiniteCycleViewPager.setOnInfiniteCyclePageTransformListener();
-
-//        horizontalInfiniteCycleViewPager.setCurrentItem(
-//                horizontalInfiniteCycleViewPager.getRealItem() + 1
-//        );
-    }
-}
