+++ BatB/app/src/main/java/com/example/batb/ImageList.java
@@ -3,6 +3,8 @@ package com.example.batb;
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
 import androidx.viewpager.widget.ViewPager;
 
 import com.example.batb.adapters.MainPagerAdapter;
@@ -10,11 +12,13 @@ import com.gigamole.navigationtabstrip.NavigationTabStrip;
 
 public class ImageList extends AppCompatActivity{
 
+    public static String[] celebName = new String[4];
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.list_img);
+        celebName[0] = "IU"; celebName[1] = "V"; celebName[2] = "Queen"; celebName[3] = "Jeong";
 
         //list
         final ViewPager viewPager = (ViewPager) findViewById(R.id.vp_main);
@@ -22,13 +26,9 @@ public class ImageList extends AppCompatActivity{
         viewPager.setOffscreenPageLimit(4);
         //tab
         final NavigationTabStrip navigationTabStrip = (NavigationTabStrip) findViewById(R.id.nts);
-        navigationTabStrip.setTitles("IU", "V", "Queen", "Jeong");
+        navigationTabStrip.setTitles(celebName[0],celebName[1], celebName[2], celebName[3]);
         navigationTabStrip.setViewPager(viewPager);
 
-    }
-
-    public interface FragmentListener {
-        void getCelebId(int a);
     }
 
 }
