+++ BatB/app/src/main/java/com/example/batb/listSplashActivity.java
@@ -1,0 +1,30 @@
+package com.example.batb;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+
+import androidx.annotation.Nullable;
+
+public class listSplashActivity extends Activity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.splash_list);
+        
+        startLoading();
+    }
+
+    private void startLoading() {
+        Handler handler = new Handler();
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent intent = new Intent(getBaseContext(), ImageList.class);
+                startActivity(intent);
+                finish();
+            }
+        }, 2000);
+    }
+}
